Abandoned Create script NOT USED
This is a script I was working on to import the Gems File GDB with corrected Field Data Types. I bailed on this when I realized we need to manipulate columns that the AlterField tool doesn't have access to (Precision and Scale). I'm just dumping it here in case there's something in it that might be useful at some point.