Implementation of Add/Edit Map Units selection. Add/Edit dialog is stubbed.
Ultimately, this combobox will be moved into a dialog, but much of the code will still be usable.

Adding a new value is clunky, because the Arc Pro ComboBox has no way to sense when a user has entered a new value. There is OnTextChanged, but this triggers for each character. Or there is this OnLostKeyboardFocus, which I've used here. It's clunky and sometimes causes ghost events. I'm not too worried about this though, because we intend to move this combobox into the add/edit dialog, at which point I can switch to a vanilla Windows combobox, which registers an Enter as a selection change.